+++ tools/src/librlink/RlinkPacketBufRcv.cpp
@@ -1,6 +1,6 @@
-// $Id: RlinkPacketBufRcv.cpp 983 2018-01-02 20:35:59Z mueller $
+// $Id: RlinkPacketBufRcv.cpp 1079 2018-12-09 10:56:59Z mueller $
 //
-// Copyright 2014-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2014-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-08  1079   1.2    use ref not ptr for RlinkPort
 // 2017-04-07   868   1.1.1  Dump(): add detail arg
 // 2017-02-19   853   1.1    use Rtime
 // 2014-12-25   621   1.0.1  Reorganize packet send/revd stats
@@ -76,10 +77,10 @@ RlinkPacketBufRcv::~RlinkPacketBufRcv()
 //------------------------------------------+-----------------------------------
 //! FIXME_docs
 
-int RlinkPacketBufRcv::ReadData(RlinkPort* port, const Rtime& timeout, 
+int RlinkPacketBufRcv::ReadData(RlinkPort& port, const Rtime& timeout, 
                                 RerrMsg& emsg)
 {
-  if (port == nullptr) 
+  if (&port==nullptr || !port.IsOpen()) 
     throw Rexception("RlinkPacketBufRcv::ReadData()", 
                      "Bad state: port not open");
   if (fRawBufDone != fRawBufSize)
@@ -89,7 +90,7 @@ int RlinkPacketBufRcv::ReadData(RlinkPort* port, const Rtime& timeout,
   fRawBufDone = 0;
   fRawBufSize = 0;
 
-  int irc = port->Read(fRawBuf, sizeof(fRawBuf), timeout, emsg);
+  int irc = port.Read(fRawBuf, sizeof(fRawBuf), timeout, emsg);
 
   if (timeout.IsZero() && irc == RlinkPort::kTout) return 0;