+++ tools/src/librlink/RlinkPacketBufRcv.hpp
@@ -1,6 +1,6 @@
-// $Id: RlinkPacketBufRcv.hpp 1052 2018-09-30 08:10:52Z mueller $
+// $Id: RlinkPacketBufRcv.hpp 1079 2018-12-09 10:56:59Z mueller $
 //
-// Copyright 2014-2017 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2014-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
 // This program is free software; you may redistribute and/or modify it under
 // the terms of the GNU General Public License as published by the Free
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-08  1079   1.2    use ref not ptr for RlinkPort
 // 2017-04-07   868   1.1.1  Dump(): add detail arg
 // 2017-02-19   853   1.1    use Rtime
 // 2014-12-25   621   1.0.1  Reorganize packet send/revd stats
@@ -40,7 +41,7 @@ namespace Retro {
                     RlinkPacketBufRcv();
                    ~RlinkPacketBufRcv();
 
-      int           ReadData(RlinkPort* port, const Rtime& timeout, 
+      int           ReadData(RlinkPort& port, const Rtime& timeout, 
                              RerrMsg& emsg);
       bool          ProcessData();
       void          AcceptPacket();