+++ tools/src/librw11/Rw11UnitDEUNA.cpp
@@ -1,4 +1,4 @@
-// $Id: Rw11UnitDEUNA.cpp 1076 2018-12-02 12:45:49Z mueller $
+// $Id: Rw11UnitDEUNA.cpp 1080 2018-12-09 20:30:33Z mueller $
 //
 // Copyright 2014-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-09  1080   1.0.2  use HasVirt(); Virt() returns ref
 // 2018-12-01  1076   1.0.1  use unique_ptr
 // 2017-01-29   847   1.0    Initial version
 // 2014-06-09   561   0.1    First draft 
@@ -72,7 +73,7 @@ void Rw11UnitDEUNA::Dump(std::ostream& os, int ind, const char* text,
 
 void Rw11UnitDEUNA::AttachDone()
 {
-  fupVirt->SetupRcvCallback(boost::bind(&Rw11CntlDEUNA::RcvCallback,
+  Virt().SetupRcvCallback(boost::bind(&Rw11CntlDEUNA::RcvCallback,
                                        &Cntl(), _1));
   Cntl().UnitSetup(0);
   return;